refactor(auth): move GitHub and OpenAI credential resolution out of agent/utils
  utils/auth.py          -> auth/resolve.py
  utils/github_token.py  -> auth/thread_token.py
  utils/github_app.py    -> auth/github_app.py
  utils/github_proxy.py  -> auth/proxy.py
  utils/openai_oauth.py  -> auth/openai_oauth.py

gateway.py stays in utils: it routes model construction through the LangSmith
LLM Gateway and is consumed by utils/model.py — reading a LangSmith key does
not make it credential resolution.

agent/auth/__init__.py stays empty. resolve.py posts auth-failure notices
through utils/slack.py and utils/linear.py while github_comments.py and
github_org_membership.py read thread_token and github_app, so the two sides
depend on each other; only submodule imports keep that from closing a cycle.
