+++ agent/dashboard/review_chat_api.py
@@ -17,11 +17,11 @@ from typing import Any
 import httpx
 from fastapi import HTTPException
 
+from agent.auth.github_app import get_github_app_installation_token
 from agent.thread_ids import reviewer_thread_id
 
 from ..review.diff import fetch_pr_diff
 from ..review.findings import REVIEWER_THREAD_KIND
-from ..utils.github_app import get_github_app_installation_token
 from ..utils.json_types import as_json_object
 from ..utils.thread_ops import langgraph_client, langgraph_url
 from .options import SUPPORTED_MODEL_IDS, canonical_model_pair, model_supports_effort
