+++ agent/middleware/sandbox_circuit_breaker.py
@@ -7,9 +7,10 @@ from typing import Any
 
 from langgraph_sdk import get_client
 
-from ..utils.github_app import get_github_app_installation_token
+from agent.auth.github_app import get_github_app_installation_token
+from agent.auth.thread_token import get_github_token
+
 from ..utils.github_comments import post_github_comment
-from ..utils.github_token import get_github_token
 from ..utils.linear import comment_on_linear_issue
 from ..utils.slack import LANGGRAPH_URL, get_active_slack_thread, post_slack_thread_reply
 from ..utils.user_messages import warning
