+++ agent/review/publish.py
@@ -25,6 +25,8 @@ from typing import Any, TypedDict
 
 import httpx
 
+from agent.auth.thread_token import GitHubAuthError
+
 from ..utils.dashboard_links import dashboard_thread_url
 from ..utils.github_checks import CheckConclusion, complete_review_check_run
 from ..utils.github_http import (
@@ -33,7 +35,6 @@ from ..utils.github_http import (
     github_client,
     github_request,
 )
-from ..utils.github_token import GitHubAuthError
 from .findings import (
     DiffSide,
     Finding,
