+++ agent/reviewer.py
@@ -39,6 +39,8 @@ from langchain.agents.middleware import ModelCallLimitMiddleware
 from langchain.agents.middleware.types import AgentMiddleware
 from langchain_core.language_models.chat_models import BaseChatModel
 
+from agent.auth.github_app import get_github_app_installation_token_with_expiry
+from agent.auth.thread_token import cache_github_token_for_thread
 from agent.sandboxes.paths import resolve_sandbox_work_dir
 from agent.sandboxes.repo_prep import materialize_trusted_skills, prepare_review_repo
 from agent.sandboxes.state import SandboxUnreachableError
@@ -115,8 +117,6 @@ from .utils import ttl_cache
 from .utils.agents_md import fetch_agents_md, fetch_scoped_agents_md
 from .utils.api_standards_skill import fetch_api_standards_skill
 from .utils.deferred_model import make_deferred_error_model
-from .utils.github_app import get_github_app_installation_token_with_expiry
-from .utils.github_token import cache_github_token_for_thread
 from .utils.model import DEFAULT_LLM_REASONING, make_model, provider_model_kwargs
 from .utils.tracing import REVIEW_TRACING_PROJECT, traced_graph_factory
