+++ agent/tools/add_finding.py
@@ -5,6 +5,8 @@ from typing import Annotated, Any
 from langgraph.config import get_config
 from langgraph.prebuilt import InjectedState
 
+from agent.auth.thread_token import get_github_token
+
 from ..review.diff import compute_diff_line_set, fetch_pr_diff, is_range_in_diff
 from ..review.findings import (
     DEFAULT_FINDING_TITLE,
@@ -22,7 +24,6 @@ from ..review.findings import (
     resolve_review_head_sha,
     thread_missing_tool_result,
 )
-from ..utils.github_token import get_github_token
 
 
 async def add_finding(
