+++ agent/tools/reply_to_finding_thread.py
@@ -2,6 +2,8 @@ from typing import Any
 
 from langgraph.config import get_config
 
+from agent.auth.thread_token import get_github_token
+
 from ..review.findings import (
     FindingInteraction,
     ReviewerThreadMissingError,
@@ -13,7 +15,6 @@ from ..review.findings import (
     update_finding_fields,
 )
 from ..review.publish import reply_to_review_comment
-from ..utils.github_token import get_github_token
 
 
 async def reply_to_finding_thread(finding_id: str, body: str) -> dict[str, Any]:
