+++ agent/utils/model.py
@@ -4,9 +4,10 @@ from typing import Any, Literal, TypedDict, Unpack, cast
 
 from langchain.chat_models import init_chat_model
 
+from agent.auth.openai_oauth import build_desktop_openai_oauth_model, desktop_openai_oauth_available
+
 from ..dashboard.options import DEFAULT_MODEL_ID, model_profile_with_context_override
 from .gateway import gateway_env_default, gateway_overrides
-from .openai_oauth import build_desktop_openai_oauth_model, desktop_openai_oauth_available
 
 OPENAI_RESPONSES_WS_BASE_URL = "wss://api.openai.com/v1"
 BASETEN_BASE_URL = "https://inference.baseten.co/v1"
