+++ agent/utils/thread_participants.py
@@ -7,12 +7,12 @@ from typing import Any
 from langgraph.config import get_config
 from langgraph_sdk import get_client
 
+from agent.auth.thread_token import get_github_token
 from agent.source_context import SourceContext
 
 from ..dashboard.agent_overrides import resolve_github_login
 from ..dashboard.user_mappings import get_mapping, login_for_email, login_for_slack_id
 from .github_comments import fetch_github_thread_participants
-from .github_token import get_github_token
 from .json_types import as_json_object, thread_metadata
 from .linear import fetch_linear_issue_participant_emails
 from .slack import fetch_slack_thread_messages
